refactor : 태그 선택 훅 분리
